+++ tests/intermediate-representation/IRGeneratorTestWithoutRegisterAllocation.cpp
@@ -70,15 +70,15 @@ TEST_F(IRGeneratorTestWithoutRegisterAllocation, VariableDeclarationStatements) 
   // <_t0, *, 2, 10> = <target, op, arg1, arg2>
   EXPECT_EQ(instructions[0]->getOperation(), Operation::MOV);
   EXPECT_STREQ(instructions[0]->getTarget()->getValue<std::string>().c_str(), "_t0");
-  EXPECT_EQ(instructions[0]->getArg1()->getValue<int>(), 2);
+  EXPECT_EQ(instructions[0]->getArg1()->getValue<int64_t>(), 2);
   EXPECT_EQ(instructions[1]->getOperation(), Operation::IMUL);
   EXPECT_STREQ(instructions[1]->getTarget()->getValue<std::string>().c_str(), "_t0");
-  EXPECT_EQ(instructions[1]->getArg1()->getValue<int>(), 10);
+  EXPECT_EQ(instructions[1]->getArg1()->getValue<int64_t>(), 10);
 
   // <_t1, +, 5, _t0>
   EXPECT_EQ(instructions[2]->getOperation(), Operation::MOV);
   EXPECT_STREQ(instructions[2]->getTarget()->getValue<std::string>().c_str(), "_t1");
-  EXPECT_EQ(instructions[2]->getArg1()->getValue<int>(), 5);
+  EXPECT_EQ(instructions[2]->getArg1()->getValue<int64_t>(), 5);
   EXPECT_EQ(instructions[3]->getOperation(), Operation::IADD);
   EXPECT_STREQ(instructions[3]->getTarget()->getValue<std::string>().c_str(), "_t1");
   EXPECT_STREQ(instructions[3]->getArg1()->getValue<std::string>().c_str(), "_t0");
@@ -91,10 +91,10 @@ TEST_F(IRGeneratorTestWithoutRegisterAllocation, VariableDeclarationStatements) 
   // <_t2, -, 7, 1>
   EXPECT_EQ(instructions[5]->getOperation(), Operation::MOV);
   EXPECT_STREQ(instructions[5]->getTarget()->getValue<std::string>().c_str(), "_t2");
-  EXPECT_EQ(instructions[5]->getArg1()->getValue<int>(), 7);
+  EXPECT_EQ(instructions[5]->getArg1()->getValue<int64_t>(), 7);
   EXPECT_EQ(instructions[6]->getOperation(), Operation::ISUB);
   EXPECT_STREQ(instructions[6]->getTarget()->getValue<std::string>().c_str(), "_t2");
-  EXPECT_EQ(instructions[6]->getArg1()->getValue<int>(), 1);
+  EXPECT_EQ(instructions[6]->getArg1()->getValue<int64_t>(), 1);
 
   // <ba, <-, _t2, null>
   EXPECT_EQ(instructions[7]->getOperation(), Operation::MOV);
@@ -104,15 +104,15 @@ TEST_F(IRGeneratorTestWithoutRegisterAllocation, VariableDeclarationStatements) 
   // <_t3, >, 6, 5>
   EXPECT_EQ(instructions[8]->getOperation(), Operation::MOV);
   EXPECT_STREQ(instructions[8]->getTarget()->getValue<std::string>().c_str(), "_t3");
-  EXPECT_EQ(instructions[8]->getArg1()->getValue<int>(), 6);
+  EXPECT_EQ(instructions[8]->getArg1()->getValue<int64_t>(), 6);
   EXPECT_EQ(instructions[9]->getOperation(), Operation::IGT);
   EXPECT_STREQ(instructions[9]->getTarget()->getValue<std::string>().c_str(), "_t3");
-  EXPECT_EQ(instructions[9]->getArg1()->getValue<int>(), 5);
+  EXPECT_EQ(instructions[9]->getArg1()->getValue<int64_t>(), 5);
 
   // <_t4, !=, 6, ba>
   EXPECT_EQ(instructions[10]->getOperation(), Operation::MOV);
   EXPECT_STREQ(instructions[10]->getTarget()->getValue<std::string>().c_str(), "_t4");
-  EXPECT_EQ(instructions[10]->getArg1()->getValue<int>(), 6);
+  EXPECT_EQ(instructions[10]->getArg1()->getValue<int64_t>(), 6);
   EXPECT_EQ(instructions[11]->getOperation(), Operation::INE);
   EXPECT_STREQ(instructions[11]->getTarget()->getValue<std::string>().c_str(), "_t4");
   EXPECT_STREQ(instructions[11]->getArg1()->getValue<std::string>().c_str(), "ba");
